[APM] Fix indefinite loading state in agent settings for unauthorized user roles (#44970)
* - handle unauthorized error to return empty list of agent settings
- display to user when a failure occurred in settings initialization
- fix incorrect settings link path

* - moved agent config index creation to the plugin setup step
- make failure text reflect a general failure to load settings

* - Rename Settings/SettingsList -> Settings/index
- Add Settings/SettingsList to render only the list/messaging
- Mention permissions issue in failure message
